Move entrypoints file to the root directory
